+++ platformio.ini
@@ -17,8 +17,9 @@ monitor_speed = 115200
 lib_deps = 
 	adafruit/Adafruit ADS1X15@^2.2.0
 	bblanchon/ArduinoJson@^6.18.3
-	ottowinter/ESPAsyncTCP-esphome@^1.2.3
+	lasselukkari/aWOT@^3.4.0
+	arduino-libraries/Ethernet@^2.0.0
 monitor_port = COM6
-upload_port = 192.168.10.151
-src_filter = +<*>  -<Ethernet_Config.h> -<Ethernet_Config.cpp>
-
+upload_port = COM6
+src_filter = +<*>  
+;lib_ldf_mode=deep
