+++ src/board.h
@@ -3,5 +3,5 @@
 #define SLOT1_CS0 8
 #define SLOT1_CS1 8
 #define ETH_RST 15
-#define ETH_CS 2
+#define ETH_CS 5
 #define led 4
